split main file into smaller files
